添加传参处理。修正abs()宏函数的正负处理行为。

+++ pak/common.h
@@ -15,7 +15,7 @@
 
 #define min(a, b) ((a) <= (b) ? (a) : (b))
 #define max(a, b) ((a) >= (b) ? (a) : (b))
-#define abs(a) ((a) >= 0 ? (a) : (-(a)))
+#define abs(a) (((int)(a)) >= 0 ? ((int)(a)) : (-((int)(a))))
 
 #define B2KB(x) ((double)(x) / 1024)
